Change Javascript to management in a separate file #21
