+++ app/javascript/packs/application.js
@@ -2,16 +2,16 @@
 // present in this directory. You're encouraged to place your actual application logic in
 // a relevant structure within app/javascript and only use these pack files to reference
 // that code so it'll be compiled.
-//= require jquery
-//= require_tree .
+//= require jquery_ujs
+
 
 import Rails from "@rails/ujs"
 import Turbolinks from "turbolinks"
 import * as ActiveStorage from "@rails/activestorage"
 import "channels"
+import "jquery"
 
 Rails.start()
 Turbolinks.start()
 ActiveStorage.start()
-
-require ('shopping/add_shopping')
+require('channels/category')
